Move all L1 processing into L1BlockHandler
